+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add `student_freeze_backbone_epochs` option to DINOv2 method to control how many epochs
+  the student backbone is frozen during training. We suggest setting it to 1 when
+  starting from DINOv2 pretrained weights. See the [DINOv2 documentation](https://docs.lightly.ai/train/stable/methods/dinov2.html)
+  for more information.
+
 ### Changed
 
 ### Deprecated